commit new code

+++ aplicacao/src/database/migrations/2022_05_24_194525_create_purchases_table.php
@@ -15,7 +15,7 @@ class CreatePurchasesTable extends Migration
     {
         Schema::create('purchases', function (Blueprint $table) {
             $table->id();
-            $table->double('discount')->nullable();
+            $table->integer('discount')->nullable();
             $table->timestamps();
         });
     }
